requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
pytz==2022.4
sqlparse==0.4.3
typing-extensions==4.4.0
